collaboration_service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.2
async-timeout==5.0.1
attrs==25.3.0
bidict==0.23.1
channels==4.0.0
channels-redis==4.2.0
dj-database-url==2.1.0
Django==5.2.6
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
frozenlist==1.7.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
msgpack==1.1.1
multidict==6.6.4
packaging==25.0
pillow==10.4.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-decouple==3.8
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.4.0
setuptools==78.1.1
simple-websocket==1.1.0
sqlparse==0.5.3
typing_extensions==4.15.0
wheel==0.45.1
wsproto==1.2.0
yarl==1.20.1